Generate a maintenance tool script specifically for the set of instances that has been created.  This tool can be used to manage the instances, check their status, start, stop and restart them, and also destroy them.  Upon destroying its instances, the tool script commits suicide.  The main motivation behind the tool script is to speed up disposal of instance sets without having to kill off all the processes manually, then delete all their respective directories.  This therefore allows multiple sets of instances to be created/destroyed without interfering with, or getting confused with, other instances.
Also add a note at the end of the instance creation to say that the tool script
is available, and also how to get help on it.  This is necessary as tool script
filenames are selected in a way that makes them unique (i.e. it won't be
overwritten when more instances are created).