Fix max_connections on chain structure and add port scan delay message.
Previously the chain structure configuration would use a value for
max_connections that was far too high and completely unnecessary.  This should
be minimal as only one instance ever connects to any other instance, so change
it to ignore the number of replicas.

Also produce a notice message after 4 seconds when scanning for free ports
as there are cases where it can take a while to find available ports without
any feedback.  Such a case is when a large set of instances have already been
set up and each port needs to be rescanned to find the end of that group of
instances.