Recheck status of instance after detecting failed start.
There are odd occasions when even after requesting a sync straight after
an attempt to start an instance, it can still report that it isn't running,
even though it is reports that immediately after.  Therefore it's worth
rechecking after 0.5 seconds to save a pointless attempt at trying to
restart it.